// src/app/models/new-requirement-model.ts
export class NewRequirement {
    reqId: any;
    jobTitle: any;
    numOfPos: any;
    department: any;
    jobType: any;
    budget: any;
    experience: any;
    openingDate: Date;
    closingDate: Date;
    location: any;
    jdFile: any;
    requestTo: any;
    aptiFile: any;
    machineTestFile: any;
    tests: any;
    notes: any;
    presets: any;
    userId: any;
}